+++ unique_toolkit/unique_toolkit/agentic/tools/utils/__init__.py
@@ -1,1 +1,19 @@
 """Utilities for tools."""
+
+from unique_toolkit.agentic.tools.utils.execution.execution import (
+    Result,
+    SafeTaskExecutor,
+    failsafe,
+    failsafe_async,
+    safe_execute,
+    safe_execute_async,
+)
+
+__all__ = [
+    "failsafe",
+    "failsafe_async",
+    "safe_execute",
+    "safe_execute_async",
+    "SafeTaskExecutor",
+    "Result",
+]