+++ unique_toolkit/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.18.0] - 2025-10-27
+- Temporary fix to rendering of sub agent responses.
+- Add config option `stop_condition` to `SubAgentToolConfig`
+- Add config option `tool_choices` to `SubAgentToolConfig`
+- Make sub agent evaluation use the name of the sub agent name if only one assessment is valid
 
 ## [1.17.3] - 2025-10-27
 - Update Hallucination check citation regex parsing pattern